requirements.txt
# requirements.txt
snowflake-connector-python
plotly-express